Version 1.4.0

+++ src/main/c/src/LLSEC_SECRET_KEY_impl.c
@@ -2,14 +2,8 @@
  * C
  *
  * Copyright 2023 MicroEJ Corp. All rights reserved.
- * This library is provided in source code for use, modification and test, subject to license terms.
- * Any modification of the source code will break MicroEJ Corp. warranties on the whole library.
+ * Use of this source code is governed by a BSD-style license that can be found with this software.
  */
-#include <LLSEC_SECRET_KEY_impl.h>
-#include <LLSEC_configuration.h>
-#include <LLSEC_mbedtls.h>
-#include <string.h>
-#include <stdint.h>
 
 /**
  * @file
@@ -18,6 +12,11 @@
  * @version @CCO_VERSION@
  * @date @CCO_DATE@
  */
+
+#include <LLSEC_SECRET_KEY_impl.h>
+#include <LLSEC_configuration.h>
+#include <LLSEC_mbedtls.h>
+#include <string.h>
 
 /**
  * @brief return the max size of the encoded key.
@@ -29,7 +28,7 @@
  * @note Throws NativeException on error.
  */
 int32_t LLSEC_SECRET_KEY_IMPL_get_encoded_max_size(int32_t native_id) {
-    LLSEC_SECRET_KEY_DEBUG_TRACE("%s (native_id = %d)\n", __func__, native_id);
+    LLSEC_SECRET_KEY_DEBUG_TRACE("%s (native_id = %d)\n", __func__, (int)native_id);
     int32_t max_size = 0;
 
     LLSEC_secret_key* secret_key = (LLSEC_secret_key*)native_id;
@@ -37,7 +36,7 @@ int32_t LLSEC_SECRET_KEY_IMPL_get_encoded_max_size(int32_t native_id) {
         max_size = secret_key->key_length;
     }
 
-    LLSEC_SECRET_KEY_DEBUG_TRACE("%s Return size = %d\n", __func__, max_size);
+    LLSEC_SECRET_KEY_DEBUG_TRACE("%s Return size = %d\n", __func__, (int)max_size);
     return max_size;
 }
 
@@ -53,13 +52,13 @@ int32_t LLSEC_SECRET_KEY_IMPL_get_encoded_max_size(int32_t native_id) {
  * @note Throws NativeException on error.
  */
 int32_t LLSEC_SECRET_KEY_IMPL_get_encoded(int32_t native_id, uint8_t* output, int32_t output_length) {
-    LLSEC_SECRET_KEY_DEBUG_TRACE("%s (native_id = %d)\n", __func__, native_id);
+    LLSEC_SECRET_KEY_DEBUG_TRACE("%s (native_id = %d)\n", __func__, (int)native_id);
 
     LLSEC_secret_key* secret_key = (LLSEC_secret_key*)native_id;
     if(NULL != secret_key) {
         (void)memcpy(output, secret_key->key, output_length);
     }
 
-    LLSEC_SECRET_KEY_DEBUG_TRACE("%s Return size = %d\n", __func__, output_length);
+    LLSEC_SECRET_KEY_DEBUG_TRACE("%s Return size = %d\n", __func__, (int)output_length);
     return output_length;
 }
